edit repair

+++ event-post-api/db/schema.rb
@@ -31,7 +31,7 @@ ActiveRecord::Schema[7.2].define(version: 2025_01_28_042406) do
   create_table "comments", force: :cascade do |t|
     t.text "content"
     t.bigint "event_id"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["event_id"], name: "index_comments_on_event_id"
@@ -48,7 +48,7 @@ ActiveRecord::Schema[7.2].define(version: 2025_01_28_042406) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "image"
-    t.integer "likes_count", default: 0, null: false
+    t.bigint "likes_count", default: 0, null: false
   end
 
   create_table "likes", force: :cascade do |t|
